[diem-framework] re-enable verifications disabled due to generic invariant
A few functions in `DiemConfig.move` are marked `verify = false` because
the global invariants are not instrumented correctly. With the recent
changes in the generic invariant handling, verification of these
functions can be re-enabeld again.

This commit also removed a spec that does not match developers'
intention: after genesis, we can still add new configurations types,
an example is the `DiemConsensusConfig` type.

Closes: #8922