[move] Fix empty source mappings for scripts
A change to use `BinaryIndexedView` in `SourceMapping` resulted in a
panic when attempting to create "dummy" source mappings for Move
bytecode representing scripts. Fix the issue by preventing the mapping
from attempting to access the script's name and address, which are only
available on modules.

In addition, add a regression test for "dummy" source mappings.

This fixes an panic in the `disassembler` when disassembling script
bytecode.

Closes: #9005