Rewrite the access control spec as two state invariants
The existing access control spec uses the schema application which is
verbose. To simplify the access control spec and make it easier to read,
this PR rewrites the existing access control spec as two state invariants
with the new spec construct `is_signer`.

This PR is mostly about cleaning-up the spec for the role-based access control
in the Diem Framework. The next step is to look into the capability-based one
(e.g., MintCapability, ...)

This PR defines `is_txn_signer`, and replaced `is_signer` with it because the
previous implementation of `is_signer` didn't fit for purpose.

Closes: #9025