[diem-framework] isolate the secondary-signer checking into a function
This part contains a loop and the loop is not very happy to be verified
repeatedly due to different instantiations of
`fun multi_agent_script_prologue<Token>`.

As a solution, this commit isolate it out into a helper function and
summarizes what the loop intends to check with `aborts_if` and `ensures`
in the helper function spec.