[move-stdlib] Standalone module for capabilities.
This adds a proposal for a generic capability module to the nursery.  This is similar designed like the Vault, but decouples the notion of capability from what it does protected (in the Vault case storage). If we agree on this design, we can refactor the Vault to be based on the new Capability module.

To read the proposal, it might be best to check out the [generated documentation](https://github.com/wrwg/diem/blob/cap/language/move-stdlib/nursery/docs/Capability.md).