[move-prover][spec-flatten] unroll function call, lambda, and block exp
This commit adds an inliner that unroll
- function calls
- lambda expressions, and
- block expressions

After the unrolling, only function arguments and the global state can be
variables in the expressions.

This serves the foundation for a more generic simplifier.

To test the effect:

```bash
cd language/move-prover/tools/spec-flatten
cargo run -- \
    ../../../diem-framework/core/sources/AccountCreationScripts.move \
    -d ../../../diem-framework/core/sources \
    -d ../../../move-stdlib/sources \
    -p inline \
    -v
```

The combine the unrolling with the triming pass, further add
`-p trim_aborts_if` after `-p inline`.