[move] Remove TextMate grammar name-access-chain
The TextMate grammar for Move, which defines how Move is highlighted
when no language server is available, is based on regular expressions.
Initially, I attempted to use a large number of regular expressions to
tweak how the following Move code was highlighted:

```
/* Example 1: */ 0xff::Name::name
/* Example 2: */ module Std::Debug { /* ... */ }
/* Example 3: */ use MyAddr::MyModule::MyStruct;
```

I thought that by detecting patterns such as
`<number-or-identifier>::<identifier>` and
`<number-or-identifier>::<identifier>::identifier`, the grammar could do
a better job highlighting each element in the name access chain. For
example, in example 3 above, the grammar can be certain that "MyModule"
refers to a module name, not a type name.

However, I now think that there are more downsides to this approach than
there are benefits.

1. Block comments can appear anywhere within these patterns:
   `0x2/**/::/**/Coin` is a valid name access chain, for example. This
   means these patterns are brittle.
2. When matching patterns with ".*::" or ".*::.*::", the tokenization
   of several elements changes as the user types "::". This can be a
   jarring experience, as identifiers that are colored as type names in
   "One::Two:" can suddenly be c:olored as module names when the user
   finishes typing "One::Two::".
3. The name access chain patterns could only slightly improve
   tokenization in certain cases, such as when used within a `use`
   declaration. But that would then in turn require regular expressions
   to match "use.*(<name-access-chain>).*;". Seeing as how block
   comments could appear within these patterns as well, the
   pattern-matching would be brittle and match or not match in
   unexpected ways.

In general I think the simpler the set of regular expressions used, the
better. The language server can instead be used to provide accurate
tokenization, once it begins to provide semantic highlighting.

Closes: #9560