[move] Add Trojan source test to TextMate grammar
Add a test that confirms the Move TextMate language grammar correctly
tokenizes comments, even when bidirectional Unicode control characters
are present.

Currently the Move source compiler does not allow for non-ASCII
characters to be included in source code -- but this test shows that
even if it did, users could be protected from this vulnerability by
using the grammar.

Closes: #9630