[move] Add messages to token errors in Move IR
Currently when the Move IR lexer or parser encounter a token they don't
expect, they exit with an error message that the token is not valid at
its location, but they don't indicate what was expected at that
location. Add a required message field to make it clearer to users.

(This falls short of modifying the Move IR lexer and parser to emit
diagnostics, which would be a much larger change, in favor of making
them slightly more user-friendly with very little changes.)

Closes: #9775