[move] Disassemble script code
9a69b3a38f5 changed how the disassembler worked for scripts: previously,
a compiled script would be transformed into a compiled module, with a
function definition that represented its `main` function. After that change,
that transformation no longer occurred, and instead code was rewritten
to operate on a "binary view," an enum that represented either a script
or a module.

However, the code change didn't account for how the disassembler worked:
the disassembler iterates over all function definitions in a module,
printing out their bytecode. Since the new code no longer created a
function definition for a script's main function, the code in the main
function was never printed.

This commit changes that, by modifying the disassembler to print out the
bytecode for a compiled script.

With this capability in place, it also migrates an additional test from
the move-ir-compiler unit test suite, to the
move-ir-compiler-transactional-tests suite.

Closes: #9941