AsMoveResolver, IntoMoveResolver
One can get a MoveResolver out of any StateView with these two traits, so there's no need to impl the resolvers everywhere (like the VersionedView and the FakeDataStore).

The returned Resolvers are wrappers around the original type (ref'ing it or owning it), impl StateView automatically, so can be used in the same way with the wrapped type.