remove MoveDbReader
Use a move resolver derived from a DbStateView in the Rest API Context,
instead of hacky impl of the resolvers on the DB directly. The Db as a
resolver doesn't provide a consistent view at a certain version and
that's not desired.

With this, all move resolvers used in the code base are implemented in
the form of RemoteStorage.

Closes: #407