Update network-identity-fullnode.md
Added an explanation, what the privat-key.txt.pub is mentioned (as a user was confused by it). Added the private-key from the sample generation to the network identity sample for consistency. Changed peerID in all samples to match the one in the sample on how to produce that ID. (I know that peerID most probably does not match the priv-key in the network identity sample, but it is the same in all samples what might be easier to read.