[executor] lazy fetch of write sets for start up
There's an edge case where when a DB is restored from a backup, the
"committed version" is deemed at the last epoch ending version because we
saved the LedgerInfo there, while there is no state snapthot before
that -- a large amount of write sets are required to compose the
StartupInfo struct as it is now.
In practise, on startup state sync will initialize things at the "synced
version" which is the latest transaction version and it should work
fine. We remove the write sets after the state checkpoint in the
StartupInfo struct and make IntoLedgerView fetch them later.

This new structure also allows up to convert the get_write_sets()
interface to get_write_set_iter() more easily -- in case we need it.

Closes: #1035