helm: add a network policy for validator rest api
We tried to install a AIT1 validator setup on GCP following the
[instructions](https://aptos.dev/tutorials/validator-node/connect-to-testnet), but the validator REST API was not exposed.

I added the following to my main.tf:

```
  helm_values = { validator: { exposeRestApi: true }}
```

But it was still not working. The reason was we were missing a network
policy for haproxy to connect to the validator on port 8080. This PR
fixes it: the API is now exposed on port 80 of the public IP assigned by
GCP.

This should also work in Azure and AWS.

PS: please consider adding our validator MIDL.dev to the
incentivized testnet program.

Closes: #1081