[api] use more correct bcs type
The goal of the original PR was to take the state from the DB and
present it to the user without any additional coding. In the ideal case,
we wouldn't even have to deserialize and then reserialize. We would just
be shipping bytes around.

The original code was re-encoding a human readable version into bcs.
That made it effectively unparseable.

The newer version does slightly better by using the types in the api
side, which are wrappers around the original types, which are also
native.

We should evaluate if for a bcs interface we want to just return back
these data types or massage it to be more of a wrapper around storage.
I'm inclined toward the former for now.