[vm] Allow certain structs as transaction parameters
This enables passing of `ASCII::String` (a struct) as a transaction parameter. We will later switch this to the new utf8 strings in the Move repo once we synced to that state.

Current restrictions (to be fixed later):

- Transaction builders do *not* yet support struct types.
- Validation would need to happen in case (as with strings) the struct has a data invariant

Implementation: refactors the code for txn arg validation in aptos-vm a little. Adds a new e2e test using the smoketest env; refactors this also a little to make it easier to write Move e2e tests.