[cli] Add command to update validator network addresses
extracted the same config from registering a new validator and used for the new command.
Seems like registering a new validator is broken on main, as bls proof of availability is missing,
but since I want to merge this to testnet branch (I assume?) that cannot be part of this diff.

how should I test this properly?