[package publishing] Add new publishing flow to cli and to smoke tests
This adds the new publishing flow to the CLI, reachable via the command `move publish --new-flow`. This specifically means deriving the `PackageMetadata` which we upload to chain programmatically from the Move package. This functionality, as well as basic package build functionality, has been factored out into a new module `move_tool/built_package.rs` (happily accepting better names).  The `BuiltPackage` is intended to become the general way how to build Move code programmatically throughout aptos-core. This PR
also applies it to the smoke tests, and adding test cases for the new flow there as well. (Next we can use `BuiltPackage` for e2e-tests in aptos-move.)

Notice that by default, `code::publish_package` stays disabled. One need to have a binary for which aptos_vm was built with the `fuzzing` feature.