[docs] Add an overview for system-integrators
System integrators have no general guide for integrating with Aptos.
They may pick up notes here and there, but nothing comprehensive that
lines up the rest of the documentation. This is an attempt to fix that.

Note, there are still several missing sections, looking for feedback and
maybe consider doing this more incrementally so that we can start
sharing this more broadly.