[smoke-tests] Add a smoke test for txn broadcasting (#2249)
### Description

Add smoke test that checks txn goes through consensus even if the local validator is not creating proposals. This behavior should be true with both mempool and quorum store.

Also improve bringing up vfn in a local swarm to:
* Avoid having a public network at validators on launch
* Avoid restart of validators on vfn startup

### Test Plan

Locally reverted #1954 and confirmed the test will fail with that bug.