[forge] Remove duplicate continuous job, increase frequency, reduce duration
Remove duplicate continuous job, instead make it the template for adding a new continuous test

Adding new continuous test is going to also require plumbing suite name through which is not set yet but I'm trying to KISS

Also reduce the continuous TPS success criteria because we want to be able to reasonably success and expectations for TPS in longer run is lower than shorter run. Not that that is noit a problem, but for now set a realistic goal for success.

Test Plan:

gh workflow run -r forgery pre-release-continuous-test.yaml