[api] expose versioned events for event specific APIs
without a version, events are of very limited use.
with a version, one can query the transaction that emitted the event and
the timestamp when that transaction was committed.
this makes it effectively possible to graph the entirety of a user
through events -- if events represent a canonical flow for the resources
they involve which should be true for Coin and Token.