[aptos-stdlib] Typed Property Tables
This is a sketch of a type-indexed map which would belong into our stdlib. This is related to PR #2673 about property maps for the token framework.

Type indexed maps have been know since long to integrate dynamic extensibility into statically typed languages. Move's global storage model uses a variation of this approach. The basic advantage is that indexing by type guarantees you get an instance of that type which satisfies all relevant invariants.