[forge] Add python forge wrapper to replace run_forge
Run forge is hard to read / test, since we have a wrapper make it maintainable.

Longer term we want to move most of this functionality to the forge cli itself, but in the meantime we shouldnt have to compile to handle certain tasks around forge execution.

Test Plan:

Running locally / forge on PR

Unittests added pass as well as #e2e