[logger] Attempt to add deduplication to logger
This is almost certainly wrong, but add the ability to deduplicate events to logger.

I was not able to think of a more clean / easy way to do this, however i think this did not work quite right since in forge now I see now logs :)

If anyone has idea why this wouldnt work and / or if this is worthwhile please comment.