[storage/api] Ensure oldest version provided to users can be queried (#2768)
* [storage/api] Ensure oldest version provided to users can be queried

There is a non-zero amount of time between requesting the oldest version
and making a second request to get that information.  This causes issues
where the data is already pruned at that point.  This should ensure that
there's at least a buffer for users.

* [storage] Add configurability to user offset window

Now you can set the amount that users would want to offset their
configs by.  This is tricky, but I just defaulted it to 20% of
the defaults

* [storage] Min viable version uses min version if pruner not enabled

* [storage] Fixup code for pruner window

* [storage] Remove dead code