[move-e2e-tests] remove archaic move serialization
Account and coin store showed how move could be serialized manually. I
already noticed that due to how easy, I mean hard, that code is to
write, there was a bug -- but it wasn't discovered because this stuff
isn't easy.

Anyway this is just bcs, so we might as well use the well defined types
rather than manually created ones. Maybe future generations would have
benefitted from seeing this, but I think they'll be happier not knowing
it existed.