[framework] add data structures for account cap offers
This exposes two new data structures for storing offers for capabilities
on the accounts. The PR intentionally only focuses on data structures
and not methods for constructing or acquiring due to
1) Methods are backwards compatible
2) Making changes to data structures is substantially more work (goldens)
3) Waiting on the PR that provides domain separation for ed25519 keys
4) Trigger early conversation

So what is one way to use this:
* An account can create at most one outstanding capability offer at a time
* To create an offer, the owner stores an address in the offer
* To claim that offer, the owner of that address calls a function, that
  function generates the appropriate capability and transfers it back

How we generate this and do pick up have some quirks. We may want
to just do this in a trivial entry point or with additional security,
such as a proof of intention -- similar like a proof key ownership.