[cli & framework] Publishing size, upgrade_policy, reactivated e2e tests
This updates the CLI for `move publish` to specify which metadata to include (none, spare, or all). There is also a CLI-time check whether the users package become too large, and what to do about it.

Furtheremore, the `upgrade_policy` is now a field of the `Move.toml` package section, as it was always planned. This lead to some cascading changes throughout. Specifically, we had to duplicate some Move test packages to represent combinations of upgrade policies.

PR also adds some missing filed for code_publishing e2e tests, and reactivates those tests.