[aptos-telemetry] Lite build information collection (#3309)
This commit introduces aptos-build-info, a library crate that collect certain build information and exposes them for consumption. The library is used in aptos-node and aptos crates to feed the aptos-telemetry crate and provide CLI version information. aptos-build-info uses shadow_rs so it is only introduced in the binary crates so as to not introduce overhead during build times. It also gets PKG_VERSION of calling crate via macros.