[aptos-rosetta] Use chain_id-block_height for block hash and remove block hash
BlockHash needs to be easy to find, and since it isn't actually used
to hash anything, this guarantees that it's unique.
This block cache is no longer necessary since we don't need to be
able to lookup hashes.  This also removes the possible out of memory
risk associated with it.