[code publishing] Optionally generate package metadata for TS SDK and other tools
As we have disabled the ModuleBundle publish path, only the new `code::publish_package_txn` can be used to publish code. This in turn requires to pass package metadata in addition to the raw bytecode. In order to enable TS SDK and other tools to get hand on this data, this PR adds new options to the compile command:

```
aptos move compile --save-metadata [ --included-artifacts=<...> ]
```

If this option is given, a file `build/<package_name>/package-metadata.bcs` will be generated which represents the metadata for this package.