[genesis] Add initial balances
This change adds the ability to define an initial list of balances to genesis ceremony. This change does the parts below the aptos cli because there are layout changes that need to accompany this.

This is needed because in mainnet we cannot simply mint ourselves token, we need to make sure they're distributed at genesis.

Test Plan: write tests? #e2e