Remove all remaining CRSN related code and comments (#3903)
### Description

Cleanup. A noop logic-wise.

CRSN is no longer in Move, and we don't have any code that actually takes the CRSN paths. It just adds complexity to the code.

This is mostly a mempool change, but also touches parts of the DB views. These should still be backwards compatible because CRSNs are never used anyway.

`AccountSequenceInfo` does not need to be an enum. For now, we keep the structure to allow future changes of sequence numbers.

### Test Plan

Existing unit and end-to-end tests.