[code publishing] Make package metadata unforgeable
This PR hardens the correctness of package metadata (arguable now completely) by allowing bytecode modules only to depend on modules which are registered via the package dependencies as listed in the metadata. If the metadata fakes the dependencies, this will be detected because the module dependencies in the bytecode are unregistered. This works in turn because the code system ensures that there is a 1:1 relation ship between package and modules (each module is part of exactly one package).

The PR also demonstrates how to design compatible mixed framework/native code upgrades. To this end, the `features` module is introduced.
For tests, the parameterized testing framework `rstest` has been instrumented to do multi-version testing. Most of those measurements are for now rather ad-hoc, and can perhaps later be generalized.

Finally, the PR corrects a wrong error code returned by the VM for init_module and for package metdata validtion. This returned VERIFICATION_ERROR
which is an internal error and leads to discard. Instead now CONSTRAINT_NOT_SATISFIED is returned which is kept. This might strictly be a breaking
change, but I'd say acceptable at this point.