[backup] Create parent directories if they dont exist
Backup cli fails if the parent dir doesnt exist for the backup, but not for the metadata.

When it fails, also it doesnt print the full path, just the filename.

This makes it create the parent dirs and also print the fullname of path if it fails.

Test Plan: run locally, see failure, with change no failure

Running fullnode locally and run the backup cli

```
RUST_BACKTRACE=1 cargo run -p backup-cli --bin db-backup -- \
	coordinator run \
	--metadata-cache-dir /opt/devnet/backup/metadata \
	--backup-service-address http://localhost:6186 \
	--state-snapshot-interval-epochs 24 \
	--transaction-batch-size 100000 \
	local-fs \
	--dir /opt/devnet/backup/data
```