[consensus] reply epoch change proof if the node is not part of validator set
This fixes a theoretic liveness problem if the only honest node in epoch moves in new epoch and doesn't
reply back the epoch change proof nor send any messages from the new epoch, the rest 2f nodes would be stuck in the
old epoch with f byzantine nodes.

This commit adds the reply mechanism back for the nodes that're not in the validator set to fix it.