[aptos-vm] disable modules during prologue
since we want to delete this as much as possible, we need to prevent
them from getting into mempool and hitting a committed but discarded
state. To that end, we need protect further upstream. We already added
an abort call in the VM for testnet to protect it, but that's a hack
considering the code situation. Moving this in here simplifies testnet,
makes sure weird behavior cannot happen on devnet (e.g., committed but
aborted ModuleBundle), and makes it so we don't have to think about it
for mainnet.