[consensus] avoid broadcasting commit vote in happy path
In happy path, we now only send the commit vote to the block proposer (still broadcast if it's a NIL block).
On receiver side, if the node aggregates the decision and is the block proposer, it'll broadcast the decision.

We still rebroadcast if there's no progress within a time frame.