[mempool] estimate in-memory size using BCS raw transaction (#4193)
### Description

The previous implementation does not work; it would just output the size of the static parts of the struct so all transactions had the same estimate. Instead we replace with BCS raw transaction size. While this does not show an exact allocation of bytes it is a good stand-in estimate -- as BCS does not attempt to compress or otherwise reduce data (e.g., RLE).

The BCS size is cached and later used to compute size on broadcast. So this only requires more work for transactions that end up not being broadcasted.

### Test Plan

Added a unit test that shows relative difference in estimated sizes. This previously failed in main because the sizes outputted were the same.