[forge][flaky] Fix consensus_stress_test
ChangingWorkingQuorumTest is extended/more complicated version of ContinuousProgressTest.
I've tested that one more, and fixed some of the bugs, which I forgot to fix in the simmpler version.

Removing simpler version, as I can just configure more complex one to skip the failures part, making them the same.

The bug relevant here - is that we cannot look at progress as substraction of rounds and subtraction of epochs separately, but as a tuple only.