[mempool] possibly fix fn_to_val_test (#4468)
### Description

The test fails and succeeds on retry in CICD sometimes: only the first transaction is propagated. There is a race condition where the broadcast from the pfn happens after txn0 is added but before txn1 is added. In that case, only txn0 is broadcasted (because the test is instrumented to only send a single message). By adding the txns first, we can avoid this.

### Test Plan

Will have to observe if flakiness goes down after land.