[smoke tests] local swarm build retries 3 times (#4461)
### Description

Adding retries on a couple tests in #4437 seems to have worked. They have not flaked on local swarm startup. So expanding to all local swarm startup.

### Test Plan

(same as #4437) Injected a retry and at least confirmed that we can successfully create a new swarm after having created a swarm (that was itself successful).