[move-e2e] remove legacy testing harnesses
These were used back in the day where we used to ensure we'd have
perfect compatibility with every version of our framework. I don't know
why we were so obsessed with compatibility to that extent.

We'll definitely want to improve on our versioning story, but that's
going to happen in a different way and in a different crate
(e2e-move-tests)