[framework] add additional data support to create_resource_account
As we started deploying our own contracts, we realized that one would
need to pass in additional data out of band to make a really standalone
contract in some cases. This isn't ideal. This example isn't
particularly elegant but it is ... generalizable. we'll benefit from a
bcs parser (not deserializer) that can be told what to parse next.