[forge] Make region per cluster, abstract cluster config
Right now region is global for a forge run which would
work fine for a single test in a single cluster but in
order to run multiple tests across multiple regions we
need to couple region to cluster.

In order to be able to store this in the config we also
have to restructure the config so that its not directly
accessing the enabled / all cluster fields so I can swap
them out safely.

Test Plan: unittests, PR