[vm] Invalidate loader cache on any module bundle loading errors
This invalidates the loader cache if an error happens either in module bundle loading or execution of `init_module`. In such case, none of the modules is committed to storage, but some may linger in the cache. This can lead to inconsistencies, like binding to `init_module` of an older (failed) version of a module.

The added test fails with the code before this PR, and succeeds with this one.