[vm] Bring back vector functions gated by feature flag
A bug in the Move compiler leads to that `0x1::vector::foo` is mapped to the actual Move function whereas only `std::vector::foo` is mapped to the new vector opcodes. Thus users which do not use `std` have deployed code on chain which still calls into the vector functions which have been recently removed.

Before we fix the compiler, we need to add support for the legacy functions. Eventually we turn this off (after all users ported) with the feature flag.

This PR does the following:

1. Adds a new feature flag NO_LEGACY_VECTOR. If this is false (default) the old vector functions are again supported.
2. Copies the vector.rs module from the Move repo into our code, and modifies it such if the above flag is set, each function will abort.

There is some larger plumbing necessary to get feature flag down to natives (which will be also useful for other applications).