[move] Updating move commit, among others fixing a bug in docgen (#5270)
This updates us to a new Move commit on the `aptos` branch pulling in [this PR](https://github.com/move-language/move/pull/624) compared to before. The changes are in the documentation generator, the prover, and the txn testing framework.

The docgen changes fix the generated overview indices for the documentation set, which now contain only the proper modules and not also the imports.