[smoke tests] test_genesis_transaction_flow: add wait checks on validator restarts (#5262)
### Description

This is one of the top flaky tests.

Observed flaky runs fail with `Waiting for nodes to catch up to version 10 timed out after 10s`. While this is probably enough time for state sync, there may be variation in node restart times, so adding the waits.

If things still fail, at least it will be clearer which is at fault.

### Test Plan

Will have to observe if the flakes go away.