[move] Fix artifact generation and test for pure Move development
This moves the test for whether artifacts are up-to-date out of framework into its own crate. This is necessary so the test is not contaminated with dependencies to `cached-packages`. Moreover, `cached-packages` is introduced as a pseudo dev-dependency to framework. This ensures that artifacts are updated if tests are called on the framework. Overall, this should smoothen out the flow for folks which only change Move code. (Albeit, extra compile time for artifacts on `cargo test -p framework`, but that is unavoidable.)