[smoke tests] use a random port instead of an OS-assigned port (with bind to 0) (#5481)
### Description

Trying to eliminate an entire class of smoke test flakiness in CICD due to conflicting ports.

Doing lsof on flaky tests in CICD, we observe that a new parallel test takes a port from a currently running node between the time it is stopped and started. When a node N goes down (by design, in the test) a port that was being used by that node N is stolen by a node M in another test. When node N restarts, the conflict occurs.

It seems very hard to hit if things were really random, yet in practice we see this at least once a day. Possibly the OS is prioritizing recently released ephemeral ports. In this PR we use pseudo-randomness, so the chances of a new process taking a port between process stop and restart should be fairly low.

(Note cargo nextest runs parallel tests in different processes, so we need to avoid conflicts between processes, so a static variable is not a solution.)

### Test Plan

Land and observe CICD for test flakiness.