[smoke tests] when getting "random" ports, use the filesystem to avoid conflicts between parallel test cases (processes) (#5421)
### Description

This should eliminate an entire class of smoke test flakiness in CICD due to conflicting ports.

Doing lsof on flaky tests in CICD, we observe that a new parallel test takes a port from a currently running node between the time it is stopped and started. When a node N goes down (by design, in the test) a port that was being used by that node N is stolen by a node M in another test. When node N restarts, the conflict occurs. (It seems very hard to hit if things were really random, yet in practice we see this at least once a day. Possibly the OS is prioritizing recently released ephemeral ports.)

cargo nextest runs parallel tests in different processes, so we need to avoid conflicts between processes. This introduces a counter file that keeps track of the last port used and a file for locking this to avoid races.

### Test Plan

Run smoke tests in CICD. Also run `cargo test` locally as a sanity check that the behavior is not changed for non-nextest uses.