[consensus] Add gracefull backoff based on chain health (#4834)
When there is high TPS, we can get in the situation where many nodes cannot catch up. If we get to very low participation rates, we are risking pauses in case of any node failures, until the rest of the nodes catch up.

Backoff starts kicking in at 80% participation, and increases if participation continues to decrease. Backoff currently is implemented via decreasing max block size.