[language] Compile time and runtime checking of view (and entry) functions
This adds the missing compile time checks for entry functions users were complaining about, and integrates this with checking for new view functions. In general, a new module `aptos_framework::extended_checks` is added which can host future checks we may want to host. Also, _generation_ of runtime metadata (error maps and new view function tags) has been moved to this module, as it is naturally related to the checks. Both are based on the move-model, a conventient way to reflect over the full content of a Move package.

It also adds some first steps towards runtime validation of view functions. This creates a new submodule `aptos_vm::verifier` in which both the previous validator for `init_module` and for transaction arguments have been moved. Here a new validator for view functions has been added which can be connected with the view function executor.