[dockerfile] Improve validator-testing image build speed (#5982)
Currently, the validator-image is built each time because the previous step invalidates the cache. This increases build time as the bcc tools should be built every single time.

This change moves the bcc tools and apt-get install steps into a separate build called validator-testing-base and builds the validator-testing image on top of this image. This however means that we need to duplicate steps from validator image to the validator-testing image, but I think we can accept this tradeoff for reduced build times.