[vm] ensure that only known attributes can be published
The original implementation allowed for someone to publish arbitrary
attributes to their module within the RuntimeModuleMetadataV1.

This makes it particularly complicated if those flags have some impact
before and after being enabled.

This PR at least makes it so that both struct and function attributes
must be known in order to be published. There will be a lot more
plumbing for the first feature that needs to be enabled at runtime...