[telemetry-svc] Ingest metrics from unknown nodes (#6405)
This commit enables the telemetry service to accept prometheus metrics from any Aptos Node, either Validator or Full Node, even if they are not whitelisted or not part of the validator set. The service will categorize any node that is able to authenticate via handshake with the service as either UnknownValidator or UnknownFullNode based on what the node identifies itself as during the authentication.

Other clean ups are included.