[telemetry-svc] Ability to uniquely identify nodes with duplicate identity (#6391)
This commit adds support to identify multiple (concurrent) runs of aptos-node with the same identity config (i.e. peer id, chain, etc). It is accomplished by generating a UUID every time a node boots up and using that UUID as part of authentication with the telemetry service. The telemetry service will include embed UUID into logs, events, and metrics so it is easily accessible.