[aptos-vm] change restriction on view functions to support legacy view (#6503)
unfortunately, some functions were compiled and published with an older
compiler on a more permissive metadata check. As a result, there are
some modules that have view function at 0 and not 1. The existing check
for view functions only allowed for them to be called if they were at 1.
This allows anything considered to be a view function to be called.