[crash] Fixes crash in access path. (#42)
This addresses https://github.com/aptos-labs/aptos-core/security/advisories/GHSA-rpw2-84hq-48jj, with a crash in validator code caused by large type tag. The resolution is to return a Result for bcs on access path and propagate this result all calls upward the stack.

No flag should be required for this fix since the alternative of not having the fix is crashing.