[helm][aptos-node] change validator/VFN AptosNet with era
Instead of bumping the Chain ID every devnet release, we can simply
change the DNS names of each of the validators based on the chain "era"
when the chain is wiped. This means that old nodes on a previous devnet
release will try to connect to an old DNS (and fail), while only new
nodes will connect to the new validators since their devnet genesis will
be up to date.