[Quorum Store] Implement end_epoch to clear state_computer between epoch end and start (#6889)
We should not be using a stale PayloadManager after epoch end, so adding a end_epoch function that sets the PayloadManager to None.

This fixes a panic that was observed because the stale PayloadManager was still held at StateComputer during sync_to called from initiate_new_epoch. The PayloadManager expects to only see commits from its epoch, which is violated if the epoch change includes multiple epochs.

changing_working_quorum_test (with failpoints) failed with panics. Rerun and observe no panics.
TODO (not in this PR): would be nice to have a test that explicitly causes a multiple epoch sync_to from epoch manager.