[Quorum Store] fix backpressure only checking on commits (#7058)
In consensus-only tests, we saw large gaps where backpressure seemed "stuck". This was because we only checked the numbers after commit, and if commit is delayed then the numbers didn't get updated at all. This PR now increments when proofs are received. It can be an overestimate until the actual cleaning takes place.

In addition
* Make the minimum time between pulls that had transactions 100 ms. 25 ms pulls seem unsustainable, even if there is no other load in the network.
* Change the config for proof-based backpressure so it is a multiple of number of validators. This makes the config more portable between networks.

Re-run consensus only tests where we saw issues.