[gha][replay-verify] fix job cancellation
Like some other expensive cron jobs, we only want to run them on the
aptos-core repo, and not anywhere else like forks. `check-aptos-core`
action only works when the repo has already been checked out, since it
uses the `gh` CLI which needs it for context. Also, make the replay jobs
depend on the `determine-test-metadata` job, so replay will not run even
if the initial check fails.

Also do the same for `module-verify`.