[model-compiler] Enabling basic type checking of Move functions (#1021)
This is a very first step to enable the move model to process Move functions from source (and not just bytecode), and thus act as a compiler.

In this initial approach, the old Move compiler is run together with the new one. This allows us to continue to have a `CompiledModule` in the `GlobalEnv`. Untangling the model from compiled module is a larger refactoring which we can factor out in different PRs.

A new model builder option `compile_via_model` is added. If set, the body of function definitions will be piped through the model's exp builder to translate expansion AST into model AST, which is no stored in the model and accessible via `FunctionEnv::get_def()`. The model AST has been extended to handle new imperative expression constructs, but the translator has not yet been extended to support imperative constructs. Test infra has been set up as well, to be refined as we proceed.

GitOrigin-RevId: 24da96d9f4b02dd9b3f2f9cb27bb93f3b5433c8a

[Prover] Added a nested loop invariant example (#961)

Added a nested loop invariant example as a test case

GitOrigin-RevId: 9a58be0da336a145b9cc90661f4360f0d7e83542

Revert "remove invariant error" (#70)

This reverts commit c93bd7a6e88aa19601707d0ef98c36d7d120ff06.

GitOrigin-RevId: 0623d65038a6ebf9e52fd904c0749e157ab3d41d

[binary format] optmize the bounds checker

GitOrigin-RevId: 5252b771084290fdd404efbc25a3d9a9d1cf604c

Fix the CFG bug from Zelig (#73)

* test pwn

* fix

GitOrigin-RevId: a99968bedd284625528279326fdd3ca9d60d0e47

add regression test (#71)

GitOrigin-RevId: ca5d1ad372c8425d0ab43aa8ab7fdbfcfe5a7bfb

Fix fmt (#74)

GitOrigin-RevId: c3f22c459df3cb2c0fee22035d813a28bb8e4918

[move] Pulling in upstream changes