[model compiler] Introduce patterns and type check assign
This introduces the new type of patterns into the model. Patterns are like expressions but represent lvalues or matches. This is used to implement type checking of assignment. Lets also have been refactored to use patterns during checking, but this isn't yet pushed into the AST, which will be a larger breaking change.

GitOrigin-RevId: 2660b7a59bb38fa65ce514a5e64f631732fe01f0