[mempool] don't remove from mempool on sequence number too new (#7676)
### Description

When execution fails with sequence number too low, this txn shouldn't be removed from mempool. This helps if the txn is too new because of consensus-internal operations, e.g., if a block execution is cut due to gas limit. Worst case, it will be unable to be ordered again until the sequence number is filled.

### Test Plan

Existing tests.