[model compiler] Refactor move-model to be independent of bytecode
This is a large refactor of the `move-model` to become independent of bytecode, and an essential step towards the new compiler. In

Originally the move-model was designed to work from bytecode as generated by the Move compiler. As such it was input the prover, docgen, and other tools. This PR makes this dependency optional:

- The model can still operate base off bytecode, supporting all exisiting use cases and tests. This is important for the incremental introduction of the new compiler infra.
- In a new execution path enabled by this PR, the model can also operate without having bytecode as input. In this case, the model builder will completely compile the model from the Move sources.

The focus of this PR is preparing the 2nd operation mode without breaking the first.

This PR also removes a number of unused crates to simplify refactoring:

- stackless bytecode interpreter
- legacy read-write set analysis
- legacy escape analysis