[TF] Add variable "bootstrap_cluster" to GCP base module
The module uses data.kubernetes_all_namespaces to dynamically get the
list of all namespaces, but that can only be done after a cluster has
been turned up.
On cluster creation, TF attempts to connect to the K8s control place
to fetch the list of namespaces, and fails.

The solution is to split cluster bootstrap in two phases. This is only
necessary while we're in the middle of the migration from PSP to
PSS.
Once all clusters have been upgraded to PSS and K8s version >=1.25,
the bootstrapping dance can be removed.