[move] Hook for VM session output format (#8189)
Move `ChangeSet` and `AccountChangeSet` types are now generic. The
motivation for this change is to avoid any particular representation being used
to communicate changes.

All Move code is fully compatible with the change, all old APIs are preserved.
However, additionally session can take a hook which dictates how resources
should be outputed. The hook is implemented as a callback which the user can
pass on the adapter's side.