[forge] Fix long running job failure
Since the migration to GCP long duration forge jobs have been failing.

I looked at this with Rustie and initially we were stumped.

The code that determines the success is based on fetching the phase of the pod that is running the job, if the phase == "succeeded" then pass skip etc or anything else fail.

When looking at the debugging output (which had auth failure) I realized that the status check could possibly have auth failure as well.

This lead me to the auth duration, which is default to 1.5hr previously in aws forge we had a longer expiration by default!

So this PR addresses this by setting the auth expiration to double the runner duration.

I also add a little error handling in case someone hits this again

Test Plan: unittest + run the test on this branch