[cached packages] Use Yaml as a format for MRB (release bundles)
This is an attempt to reduce merge conflicts in the checked-in head.mrb file in `cached-packages`, using Yaml instead of BCS. Yaml should have a drastically better merge behavior. There is a 10x price in size to pay for .mrb, buts is yet for all our framework packages below 10mb.