[compiler v2] Integration into package system and Aptos CLI
This integrates the v2 compiler into the `move-package` code and makes it available to the `aptos` CLI. One can now compiler with the v2 compiler as in

```
aptos move compile --compiler-version=v2
```

However, compilation of already the Move stdlib fails because of #9330 (effecting bit operations in `fixed_point32.move`) and #9696 (effecting inline functions with lambdas in `vector.move`). Fixing those issues does not fit in this PR which is by itself not trivial.

Implementation wise, since the package system is build on top of internals of the v1 compiler, we provide a new v2 driver function parameter to the general entrypoint `compile_with_driver`. We also add a new flag for the compiler version which is plumbed the many layers of configurations from Aptos CLI down to Move compiler.

This PR also disables the use of the bytecode produced by v2 for production, by assigning a new experimental version number to it which is only allowed in testing mode.

This is tested manually.