[forge] Split duration more usefully, for more reliable stats
We have a sweep of tests, with some being in 1 phase and some in 2.
Previously, each test would get ~5 minutes, giving it 2.5 minutes per phase.

With this, we make all phases equal, i.e. ~4 minutes, so we get similarly stable results
for each phase, as we get for full runs, without needing to extend runtime.