[e2e-move-tests] Cache module building across tests
We have multiple tests using the same move code in data folders, causing them to be rebuilt over and over.
Adding to harness methods that cache building, so it can be reused.

Haven't changed the default methods to do the caching, as caller can decide to modify the files,
and then rebuilding is needed - so it is up to the caller to decide if it wants caching or not.

This speeds up e2e-move-tests on my machine from 90s to 50s.