[marketplace-example] Fix royalties edge conditions
There were two bugs with royalties and listings.  One was that in
v1, royalties could have a denominator 0, or be greater than 100%
in legacy NFTs.  This means that it's possible that payouts don't
work correctly, or take more than expected.  Now, royalties are
bounded to 0-100%.

The second issue was that commission was taken after royalties,
but didn't consider that royalties could be 100%.  Now, royalties
are taken first, and commission is taken out of the remainder.
This does mean the marketplace may not have any commission if the
royalties are set to 100%.