[Executor] Refactoring and Optimization of the Block STM (#10081)
Removed CommitGuard in the executor - using the standard scopeguard
Any thread can be a coordinator (The term is not used any more). Attempting to coordinate only after a successful validation. Minor optimizations
Parallel processing of outputs
Fixed a potential lost wake-up
Additionally, for notifications (e.g. for sharding) this commit ensures that if a transaction can be committed after validation it will get committed there may not be "fake" abort notifications (due to an earlier skiprest).