[compiler v2] Fixing bug in spec block analysis
This closes #10885. The fix is a bit more involved as thought because the way how inlined spec blocks are type checked needed to be changed:

- Inlined spec blocks _cannot_ be checked as they appear in the code, since they are not supposed to influence type inference, and since they are designed to work in their own inference context.
- Rather the entire function body needs to be checked first, and then subsequently the spec blocks, each with the context of locals in which they are defined, at a point where the types of the locals have been inferred.

This PR implements this new scheme by collecting spec blocks and only at the end of a function body, post-processes them.

Adds the test from the bug report.