[compiler-v2] v1-v2 test inventory tooling and data
This adds a little program (in `move-compiler-v2/tools/testdiff`) to make an inventory and record it of the mapping of v1 and v2 tests. This program is then be used by a script in `move-compiler-v2/tests/update_v1_diff.sh` to updated stored information about:

- Any unmatched tests, that is tests which appear in the v1 tree but have no "match" in the v2 (see below of what is meant with match)
- Any matched tests, as pairs of file names line-by-line of there `.exp` files. This maybe used for later processing with diff tools.

This info is in two files which are intended to be checked into the repo.

The rules for matching take into account that tests have been ported directory by directory from v1 to v2. If for example in v1 there is `move_check/typing`, the directory `typing` will appear somewhere in the v2 test hierarchy, dependent on to which phase the test belongs. (In this case, it is the type checking tests in v2, thus `checking/typing`.) There is one exception currently (which we may want to correct in a later PR) with `reference-safety`, but the code can easily handle such exceptions.  Given this logic we can identify tests as the pair of their parent directory (`typing`) and a given test in there (`some_test.move`), and search for matching pairs in the directory hierarchy.