[compiler-v2] Fixes a bug in borrow graph shrinking (#11250)
* [compiler-v2] Fixes a bug in borrow graph shrinking

Closes #11221.

When a temporary goes out of scope, nodes related to that temporary are removed from the graph, provided they are not borrowed. This process recursively continues, visiting nodes representing other temporaries, cleaning up large parts of the graph. There was a significant bug in this cleanup, namely ignoring whether a node and a related temporary are also referenced from the outside. This is for example the case if the temporary is used later in the program. It appears this bug only hits if different control flows contribute to the graph, as is in the test case `return_borrowed.move`.

* Addressing reviewer comments.