[Quorum store] fix sender batch size check (#11292)
### Description

Sender side was not checking for batch max bytes, causing batch to be rejected on the receiver side. Adding the checks.

With default settings and prod txn size, we didn't run into this issue, but it was found in forge with large transactions.

We also change the mempool txn byte limit from checking the raw txn's length to the signed txn's length, as this is what is used for the batch bytes check, and also what is sent over the wire.

### Test Plan

Added unit test.