[movevm] more validation for sponsored transaction
Test if the account resource exists before attempting to create the
account resource. This solves potential devx issues around gas fees and
also prevents potential errors when the account exists but sequence
number 0 isn't expensive enough.

added a new test for this condition that failed prior to the fix and
also retrofitted another test to be more aware of this corner case