[compiler-v2] Fixes an issue in post-processing of exp_builder
During building of expressions, placeholders are used for expressions which can't be decided because of incomplete type inference. At the end when inference is complete, a post processing phase is run to replace those placeholders. This uses the expression rewriter, but that one doesn't descend on rewrite steps.

This adds a new `RewrittenAndDescend` variant to rewrite functions which instructs the rewriter to descend on rewritten exps. Fortunately, this is cleanly possible because we not longer use `Result` as a return value.

Adds one test case which exposed this behavior from `simple_map` in the stdlib.