ristretto
Demonstrates in test `destroy_after_call` that a mutable reference is destroyed to late. There are two versions of the test, one in `explicit-drop` unit tests, the other in transactional `misc` tests. For the first one an existing mut ref in $t3 is destroyed at code offset 7, however, before this at offset 6 an imut ref is created.