[compiler v2] Release locals which are introduced but die in the same instruction (#11511)
* [compiler-v2] Only aggressively copy if type has drop ability

The hack to work around #11223 in live var analysis may cause ability errors if the value does not have the drop ability. This is because if we copy to aggressively, we may need to destroy values later, and this requires those to have a drop ability. This change fixes this by requiring to have drop to apply the hack.

* Enable the fix, demonstrating the delta

* ristretto

Demonstrates in test `destroy_after_call` that a mutable reference is destroyed to late. There are two versions of the test, one in `explicit-drop` unit tests, the other in transactional `misc` tests. For the first one an existing mut ref in $t3 is destroyed at code offset 7, however, before this at offset 6 an imut ref is created.

* [compiler v2] Release locals which are introduced but die in the same instruction

Reference analysis did not correctly release temps which are assigned and die inn the same instruction.

The first commit shows the issue, the 2nd the solution. See also individual commit messages.

This also modifies the dissassembler to print the function definition index for easier debugging of failures.

* remove debugging output

* Addressing reviewer comments

---------

Co-authored-by: Brian R. Murphy <132495859+brmataptos@users.noreply.github.com>