[compiler-v2] Adding logging to the compiler
Until now we did not had a systematic way how to log information for debugging purposes. This uses the popular [`log`](https://crates.io/crates/log) facade as a basic API for logging. The macros `error!`, `warn!`, `info!`, `debug!`, and `trace!` can be used like `println!`.  The backend we are installing for the compiler is configured via an an env var and described in the header of the new `logging.rs` module. See there for more discussion.

A few first log instructions are also inserted in the code.