[compiler v2] Fixing log specification file separator char (#11874)
A module path for log filters has the form `crate::mod1::mod2`, but `:` was also used as a file separator, as in `MVC_LOG=info:test.log`. Replaced this with `MVC_LOG=info@test.log`. Also the string is now split on `@` from the right, so if there are multiple of those chairs, we only take the last segment as file name.