[compiler v2] New Borrow Analysis
This is a major rewrite of the reference safety analysis. This PR closes #11251 which shows the problem with the previous version: essentially, it is necessary to allow for a sequence of instructions an "unsafe" borrow graph and only enforce safety at particular instruction boundaries. The simplest example to demonstrate the problem is `f(&mut s.x, &mut s.y)`. If broken down to instructions, at some point there will be a reference to `&mut s` where at the same time there is another one `&mut s.f`, which is unsafe if those refs would be passed as function parameters. The rewrite does the following:

- Instead of checking safety of borrow operations at the time they are performed, the graph is now constructed first, and later checked for safety at particular instructions (read ref, write ref, and function calls)
- The borrow graph stays mostly the same, instead of Skip edges for graph glueing which became overly complex, glueing is now performed by graph renaming.
- There are various enhancements in error reporting (and potentially because of the delayed borrow checking also some regressions)

It might be useful to review the new version without diff from scratch.