[api] add a long poll API /transactions/wait_by_hash/:txn_hash (#12419)
### Description

The motivation is to reduce e2e latencies by learning of committed responses immediately. Currently, it is up to the client to poll /transactions/by_hash and there is overhead on the poll intervals and the RTT of the poll.

There is concern around holding large numbers of connections for an extended time -- to alleviate this we put a limit on active long polls, and simply return immediately when that limit is reached. The API does not give guarantees about how long the long poll will wait.

### Test Plan

Added unit tests. Deployed locally and in previewnet and tested long poll behavior.