[compiler-v2] Finishing refactoring of abilities + type checker overhaul
This concludes moving all ability checking into the type inference algorithm by representing checks as type constraints. It also introduces some new constraints to deal with some other checks which have done previously ad-hoc (whether types are allowed to be references, tuples, phantom).

The PR turned out to be a larger refactoring of the checker as expected. On the good side, a lot of redundant ad-hoc code could be removed and replaced by more generic code.

Closes #12437 and closes #12674