[compiler-v2] A simple optimization reducing unnecessary temp creation (#12720)
I came along this while investigating the freeze bug, but it is worth its own PR because a lot of baselines are changed.

Before this PR, when we had `let x = e; s`, the bytecode generator create a useless temporary and generate code like this `t := e; x := t; s`. Now it detects the special case of a simple variable let and directly generates `x := e; s`.

Consequently, this leads to a lot of baseline changes as we now generate less code. All semantic (transactional, move-unit, e2e) pass with this. Note that the change wouldn't work before Vineeth landed #12586. In fact the old treatment masked more instances of the bg fixed in that PR because the unnecessary temps avoided it.

Notice that coalescing may have elminated some of those temps but not in all cases because of over-approximation of borrowing. Also it appears that borrow semantics in the byetcode verifier is unfortunately sensitive regards duplication of values in temps (that's why I came along this when looking at the freeze bug), so we want to avoid those unneeded temps for the case we compile without optimizations.