Fix #12958: tests being run from library code when aptos-move/move-examples are run with `MOVE_COMPILER_V2=true cargo test`.
Change compiler-v2 sources and dependencies from `Vec<String>`
filenames to `Vec<PackagePaths>` full package descriptors.
This attaches a package name to input files, allowing a
filter to limit unit tests to the current package.  Since
these are bigger structures, we pass `&` instead of object.

There is some messiness due to `PackagePaths` really being
a generic type with default instantiation using
`move_symbol_pool::Symbol` which isn't common in the
move-model code.

Also, it turns out that disassembler uses script module
names whose name mangling has changed, so there is a
change to `move-cli/src/base/disassemble.rs`, along with
a more verbose error (listing the available functions)
in the case of a failure to find the requested module,
which was useful debugging and was left in.

Since related code was touched in EVM-related code in
`third_party/move/tools/move-unit-test/src/test_runner.rs`
I fixed it to compile but there is still an issue with
the `named_address_map` which I haven't debugged.
(See Issue #13119.)