[quorum store] better handling of proofs in proof coordinator (#13163)
## Description
The motivation is our metrics show failed quorums when there are late votes, even if the quorum was in fact reached before the late votes. There were subtle bugs in metrics when inline batches are used (before getting quorum) and/or batches with the same digests (either remotely or locally) are created.

Thinking a bit more, there's no reason to use digest in proof coordinator, so changing to use the batch info instead. This means we no longer have to disambiguate between batches that are possibly created with the same digests (either remotely or locally). This simplifies the code significantly while dealing with the above issues.

## How Has This Been Tested?

Ran forge stable workload sweep test, which has gas cut