Fix SMT proof fetching for lifted leaf
In the presense of deletion, a leaf node on the SMT can be lifted, so
the hash value won't be matching that in the DB.

Examining more, it's obvious that in such cases proof fetching is not
needed in the first place.

Original thought was in these rare cases we just waste a constructed
Child to make the code simply, however it wasn't considered that there's
a panic on hash_cache miss. I think the panic is valuable assertion so
this is fixing it by some extra code.