flatten json log fields in indexer logs.
This will change emitted logs from:
{ .... fields: { foo: "bar", bla: "baz" } }
to
{ foo: "bar", bla: "baz" }

i.e. it will put all "fields" emitted into the root of the json object. We do more or less the same in api-gateway and kind of the rest of aptos-core which makes querying emitted logs a bit easier/intuitive as you don't need to prefix every single query field with `fields.` when querying.