[move] Update `move_pr.sh` to run more tests
This updates the `move_pr.sh` script to run the tests also covered by #13446 as integration tests, namely all kind of tests depending on the MOVE_COMPILER_V2 env var.

With `move_pr.sh -i2` both integration tests without the MOVE_COMPILER_V2 flag are run (the `-i` option) as with the flag set (`-2`). This mechanism was already before there, only the crate lists have been updated.

There are some `.v2_exp` files which did not exists or where out-of-date which are also updated with this PR.

We currently need a workaround to avoid `aptos-cached-packages` is build when MOVE_COMPILER_V2 is set, as this will generate new .md files for the framework based on some bug fixes the v2 compiler has, that appear as changes in the client. The script ensures that this crate is build before switching the env var to on.

As a sideffect, this also removes the 2nd dependency from RUST_BACKTRAC in exp files, in `move-compiler/src/diagnoses`. The same mechanism as already in place in `model.rs` is used: in order to get backtrace into error output, one must set both `RUST_BACKTRACE` and `MVC_BACKTRACE`.