[compiler-v2] Pattern matching for struct variants (aka enum types)
This implements the Move 2 `match` expression on stackless bytecode level. The PR consists of the following parts:

1. *Extending the Bytecode*

There are the following new bytecode operations:

```
dest := TestVariant(struct_id, variant, src)
dest := PackVariant(struct_id, variant, srcs)
dests := UnpackVariant(struct_id, variant, src)
dest := BorrowFieldVariant(struct_id, variant, field, src)
```

The `TestVariant` and `BorrowFieldVariant` operations  expect references to the variant struct. The
`UnpackVariant` and `BorrowFieldVariant` operations abort if the `src` value is not a value of
the given variant.

Currently there is no attempt made at optimizing branch cascades using `TestVariant` by some kind of switch
instruction. Notably, this simplifies the implementation as there also no new control flow operations,
and the remaining part of the stackless bytecode framework should operate without changes.

2. *Translating Matches*

Matches are translated into cascades of test & branch instructions. The translation is complicated by
the need for 'probing' whether a value can be matched before it is consumed. See for discussion in the
code how this problem is solved.

3. *Checking Match Coverage*

Even though we generate robust match code which is resilient against addition of new match variants
by explicit `abort` if there is no match, at compile time complete match coverage is required.
Checking for this is non-trivial and documented in the code.

4. *Other*

There are some only indirectlyu related changes in this PR. To implement non-destructive matches
with conditions, it appeared handy to have `*&x` to be equivalent to `x`. This seems to be general
useful and is AFAICT also in Rust.