[api] print SingleSender signature / key variants
Poem's display functionality is a bit limiting and doesn't seem to
handle printing enums effectively, so one needs to leverage a wrapping
struct to effectively display a type and value field.

This PR introduces just that.

This means that anyone parsing single or multi key sender json output
will be broken, of course, the data was largely useless to begin with.

Other aspects remain untouched so this is a minimally breaking changes
PR.