Let BorrowVariantField carry a set of variants
Figured that the implementation of commmon field selection had a bug, and worked the fix into this PR to have right away the correct opcodes in the VM.

The problem is that even if an enum has a set of common fields, this is not guaranteed to stay like this at runtime, since our enums are open-ended at runtime. That is, new variants can be added without code being recompiled. Therefore, if we have a common field `x` between variants, we cannot just emit a `BorrowField(common_offset)`, but we must check at runtime whether the variants with the common field are actually present, thus need to emit something like `BorrowVariantField(V1|V2|..|Vn, common_offset)`. This leads to changes in the handle representation and tables in the VM, which I wanted to have manifested before starting working on the interpreter.

The compilation also needed to be changed at various places to make this happen.