Restore crates from upstream
When syncing with upstream, merge conflicts of the "changed by them/
removed by us" kind have been resolved by updating the conflicting
files to the latest upstream revision. This only partially recovered
the source for crates that have been removed for Monza.
Restore the crates fully.

Some of the changes made for Monza are overwritten from upstream
to reduce the delta and allow resurrecting some crates erased
in the work for Monza:
- ring dependency removal from aptos-crypto, moving the noise
  implementation to aptos-network;
- removal of tokio synchronization slots from aptos-types.

This means this branch no longer supports Monza, and those changes
would need to be re-applied if the Monza direction is taken again.